+++ src/resolvers/question.ts
@@ -81,6 +81,7 @@ export const resolvers: Resolvers = {
 
           // create the new question
           const newQuestion = await question.create(context);
+
           if (!newQuestion?.id) {
             // A null was returned so add a generic error and return it
             if (!question.errors['general']) {